refactor: Project Section 타입 전부 분리완료 (type은 interface로 변경)
