+++ front/src/app/project/_components/ProjectDescription.tsx
@@ -1,9 +1,5 @@
+import { ProjectDescriptionSectionProps } from "@/type/project/ProjectDescriptSection-type";
 import React from "react";
-
-interface ProjectDescriptionSectionProps {
-  title: string;
-  paragraphs: string[];
-}
 
 const ProjectDescription = ({
   title,
