+++ front/src/app/project/_components/ProjectDetails.tsx
@@ -1,17 +1,8 @@
 "use client";
 
-import { ProjectSection } from "@/type/ProjerctSection-type";
+import { ProjectDetailsSectionProps } from "@/type/project/ProjectDetails-type";
 import Image from "next/image";
 import React from "react";
-
-type ProjectDetailsSectionProps = {
-  title: string;
-  sections: ProjectSection[];
-  imageSize?: {
-    width: number;
-    height: number;
-  };
-};
 
 const ProjectDetails = ({
   title,
