+++ front/src/type/project/ProjectDescriptSection-type.ts
@@ -1,0 +1,4 @@
+export interface ProjectDescriptionSectionProps {
+  title: string;
+  paragraphs: string[];
+}
