+++ front/src/type/project/ProjectDetails-type.ts
@@ -1,0 +1,10 @@
+import { ProjectSection } from "./ProjectSection-type";
+
+export interface ProjectDetailsSectionProps {
+  title: string;
+  sections: ProjectSection[];
+  imageSize?: {
+    width: number;
+    height: number;
+  };
+}
