+++ front/src/type/project/ProjectRetrospective-type.ts
@@ -1,0 +1,4 @@
+export interface Retrospective {
+  learned: string;
+  regret: string;
+}
