+++ front/src/type/project/ProjectTechStack-type.ts
@@ -1,0 +1,11 @@
+export interface TechItem {
+  category: string;
+  tech: string;
+  reason: string;
+}
+
+export interface TechStackTableProps {
+  title: string;
+  description?: string;
+  items: TechItem[];
+}
