+++ front/src/type/project/ProjectUserFlow-type.ts
@@ -1,0 +1,4 @@
+export interface UserFlowSectionProps {
+  title?: string;
+  images: { src: string; alt: string; width: number }[];
+}
